requirements.txt
flask
requests
Pillow
simplejson